refactor node.rs
