+++ components/script/dom/node/mod.rs
@@ -3,9 +3,16 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 pub(crate) use self::node::*;
+pub(crate) mod children_mutation;
+pub(crate) mod context;
 pub(crate) mod focus;
 pub(crate) mod iterators;
+pub(crate) mod layout;
+pub(crate) mod methods;
 #[allow(clippy::module_inception, reason = "The interface name is node")]
 pub(crate) mod node;
 pub(crate) mod nodeiterator;
 pub(crate) mod nodelist;
+
+pub(crate) use children_mutation::ChildrenMutation;
+pub(crate) use context::{BindContext, IsShadowTree, MoveContext, UnbindContext};
