+++ components/script/dom/node/node.rs
@@ -4,7 +4,6 @@
 
 //! The core DOM types. Defines the basic DOM hierarchy as well as all the HTML elements.
 
-use std::borrow::Cow;
 use std::cell::{Cell, LazyCell, UnsafeCell};
 use std::cmp::Ordering;
 use std::default::Default;
@@ -27,23 +26,19 @@ use js::jsapi::JSObject;
 use js::rust::HandleObject;
 use keyboard_types::Modifiers;
 use layout_api::{
-    AxesOverflow, BoxAreaType, CSSPixelRectVec, GenericLayoutData, HTMLCanvasData, HTMLMediaData,
-    LayoutElementType, LayoutNodeType, NodeRenderingType, PhysicalSides, SVGElementData,
-    SharedSelection, TrustedNodeAddress, with_layout_state,
+    AxesOverflow, BoxAreaType, CSSPixelRectVec, GenericLayoutData, NodeRenderingType,
+    PhysicalSides, TrustedNodeAddress, with_layout_state,
 };
 use libc::{self, c_void, uintptr_t};
 use malloc_size_of::{MallocSizeOf, MallocSizeOfOps};
-use net_traits::image_cache::Image;
-use pixels::ImageMetadata;
 use script_bindings::cell::{DomRefCell, Ref, RefMut};
 use script_bindings::codegen::GenericBindings::EventBinding::EventMethods;
 use script_bindings::codegen::InheritTypes::DocumentFragmentTypeId;
 use script_bindings::reflector::{DomObject, DomObjectWrap, reflect_dom_object_with_proto_and_cx};
 use script_traits::DocumentActivity;
 use servo_arc::Arc as ServoArc;
-use servo_base::id::{BrowsingContextId, PipelineId};
+use servo_base::id::PipelineId;
 use servo_config::pref;
-use servo_url::ServoUrl;
 use smallvec::SmallVec;
 use style::Atom;
 use style::context::QuirksMode;
@@ -57,19 +52,18 @@ use xml5ever::{local_name, serialize as xml_serialize};
 
 use crate::conversions::Convert;
 use crate::document_loader::DocumentLoader;
+use crate::dom::ChildrenMutation;
 use crate::dom::attr::Attr;
 use crate::dom::bindings::codegen::Bindings::AttrBinding::AttrMethods;
 use crate::dom::bindings::codegen::Bindings::CSSStyleDeclarationBinding::CSSStyleDeclarationMethods;
 use crate::dom::bindings::codegen::Bindings::CharacterDataBinding::CharacterDataMethods;
 use crate::dom::bindings::codegen::Bindings::DocumentBinding::DocumentMethods;
-use crate::dom::bindings::codegen::Bindings::ElementBinding::ElementMethods;
 use crate::dom::bindings::codegen::Bindings::HTMLCollectionBinding::HTMLCollectionMethods;
 use crate::dom::bindings::codegen::Bindings::HTMLElementBinding::HTMLElementMethods;
 use crate::dom::bindings::codegen::Bindings::NodeBinding::{
     GetRootNodeOptions, NodeConstants, NodeMethods,
 };
 use crate::dom::bindings::codegen::Bindings::NodeListBinding::NodeListMethods;
-use crate::dom::bindings::codegen::Bindings::ProcessingInstructionBinding::ProcessingInstructionMethods;
 use crate::dom::bindings::codegen::Bindings::ShadowRootBinding::ShadowRoot_Binding::ShadowRootMethods;
 use crate::dom::bindings::codegen::Bindings::ShadowRootBinding::{
     ShadowRootMode, SlotAssignmentMode,
@@ -77,17 +71,16 @@ use crate::dom::bindings::codegen::Bindings::ShadowRootBinding::{
 use crate::dom::bindings::codegen::Bindings::WindowBinding::WindowMethods;
 use crate::dom::bindings::codegen::UnionTypes::NodeOrString;
 use crate::dom::bindings::conversions::{self, DerivedFrom};
-use crate::dom::bindings::domname::namespace_from_domstring;
 use crate::dom::bindings::error::{Error, ErrorResult, Fallible};
 use crate::dom::bindings::inheritance::{
-    Castable, CharacterDataTypeId, ElementTypeId, EventTargetTypeId, HTMLElementTypeId, NodeTypeId,
-    SVGElementTypeId, SVGGraphicsElementTypeId, TextTypeId,
+    Castable, CharacterDataTypeId, EventTargetTypeId, NodeTypeId,
 };
 use crate::dom::bindings::root::{
     Dom, DomRoot, DomSlice, LayoutDom, MutNullableDom, ToLayout, UnrootedDom,
 };
 use crate::dom::bindings::str::{DOMString, USVString};
 use crate::dom::characterdata::CharacterData;
+use crate::dom::context::{BindContext, IsShadowTree, MoveContext, UnbindContext};
 use crate::dom::css::cssstylesheet::CSSStyleSheet;
 use crate::dom::css::stylesheetlist::StyleSheetListOwner;
 use crate::dom::customelementregistry::{
@@ -100,17 +93,11 @@ use crate::dom::element::{CustomElementCreationMode, Element, ElementCreator};
 use crate::dom::event::{Event, EventBubbles, EventCancelable, EventFlags};
 use crate::dom::eventtarget::EventTarget;
 use crate::dom::globalscope::GlobalScope;
-use crate::dom::html::htmlcanvaselement::HTMLCanvasElement;
 use crate::dom::html::htmlcollection::HTMLCollection;
 use crate::dom::html::htmlelement::HTMLElement;
-use crate::dom::html::htmliframeelement::HTMLIFrameElement;
-use crate::dom::html::htmlimageelement::HTMLImageElement;
 use crate::dom::html::htmllinkelement::HTMLLinkElement;
 use crate::dom::html::htmlslotelement::{HTMLSlotElement, Slottable};
 use crate::dom::html::htmlstyleelement::HTMLStyleElement;
-use crate::dom::html::htmltextareaelement::HTMLTextAreaElement;
-use crate::dom::html::htmlvideoelement::HTMLVideoElement;
-use crate::dom::html::input_element::HTMLInputElement;
 use crate::dom::iterators::ShadowIncluding;
 use crate::dom::mutationobserver::{Mutation, MutationObserver, RegisteredObserver};
 use crate::dom::node::iterators::{
@@ -119,15 +106,13 @@ use crate::dom::node::iterators::{
 };
 use crate::dom::node::nodelist::NodeList;
 use crate::dom::pointerevent::{PointerEvent, PointerId};
-use crate::dom::processinginstruction::ProcessingInstruction;
 use crate::dom::range::WeakRangeVec;
 use crate::dom::raredata::NodeRareData;
 use crate::dom::servoparser::html::HtmlSerialize;
 use crate::dom::servoparser::serialize_html_fragment;
 use crate::dom::shadowroot::{IsUserAgentWidget, ShadowRoot};
-use crate::dom::svg::svgsvgelement::SVGSVGElement;
 use crate::dom::text::Text;
-use crate::dom::types::{CDATASection, KeyboardEvent};
+use crate::dom::types::KeyboardEvent;
 use crate::dom::virtualmethods::{VirtualMethods, vtable_for};
 use crate::dom::window::Window;
 use crate::layout_dom::{ServoDangerousStyleElement, ServoDangerousStyleNode};
@@ -250,7 +235,7 @@ bitflags! {
 /// <https://dom.spec.whatwg.org/#concept-node-insert>
 /// <https://dom.spec.whatwg.org/#concept-node-remove>
 #[derive(Clone, Copy, MallocSizeOf)]
-enum SuppressObserver {
+pub(super) enum SuppressObserver {
     Suppressed,
     Unsuppressed,
 }
@@ -261,6 +246,43 @@ pub(crate) enum ForceSlottableNodeReconciliation {
 }
 
 impl Node {
+    // Getters for internal values
+    pub(super) fn parent_node(&self) -> &MutNullableDom<Node> {
+        &self.parent_node
+    }
+
+    pub(super) fn first_child(&self) -> &MutNullableDom<Node> {
+        &self.first_child
+    }
+
+    pub(super) fn last_child(&self) -> &MutNullableDom<Node> {
+        &self.last_child
+    }
+
+    pub(super) fn next_sibling(&self) -> &MutNullableDom<Node> {
+        &self.next_sibling
+    }
+
+    pub(super) fn prev_sibling(&self) -> &MutNullableDom<Node> {
+        &self.prev_sibling
+    }
+
+    pub(super) fn get_owner_doc(&self) -> &MutNullableDom<Document> {
+        &self.owner_doc
+    }
+
+    pub(super) fn get_rare_data(&self) -> &DomRefCell<Option<Box<NodeRareData>>> {
+        &self.rare_data
+    }
+
+    pub(super) fn flags(&self) -> &Cell<NodeFlags> {
+        &self.flags
+    }
+
+    pub(super) fn layout_data(&self) -> &DomRefCell<Option<Box<GenericLayoutData>>> {
+        &self.layout_data
+    }
+
     /// Adds a new child to the end of this node's list of children.
     ///
     /// Fails unless `new_child` is disconnected from the tree.
@@ -643,14 +665,8 @@ impl Node {
             NodeRenderingType::Rendered
         )
     }
-}
 
-impl Node {
-    fn rare_data(&self) -> Ref<'_, Option<Box<NodeRareData>>> {
-        self.rare_data.borrow()
-    }
-
-    fn ensure_rare_data(&self) -> RefMut<'_, Box<NodeRareData>> {
+    pub(super) fn ensure_rare_data(&self) -> RefMut<'_, Box<NodeRareData>> {
         let mut rare_data = self.rare_data.borrow_mut();
         if rare_data.is_none() {
             *rare_data = Some(Default::default());
@@ -801,7 +817,8 @@ impl Node {
     }
 
     pub(crate) fn ranges_is_empty(&self) -> bool {
-        self.rare_data()
+        self.rare_data
+            .borrow()
             .as_ref()
             .is_none_or(|data| data.ranges.is_empty())
     }
@@ -1654,7 +1671,8 @@ impl Node {
     }
 
     pub(crate) fn containing_shadow_root(&self) -> Option<DomRoot<ShadowRoot>> {
-        self.rare_data()
+        self.rare_data
+            .borrow()
             .as_ref()?
             .containing_shadow_root
             .as_ref()
@@ -1716,7 +1734,7 @@ impl Node {
 
     /// Returns the node's `unique_id` if it has been computed before and `None` otherwise.
     pub(crate) fn unique_id_if_already_present(&self) -> Option<String> {
-        Ref::filter_map(self.rare_data(), |rare_data| {
+        Ref::filter_map(self.rare_data.borrow(), |rare_data| {
             rare_data
                 .as_ref()
                 .and_then(|rare_data| rare_data.unique_id.as_ref())
@@ -2114,309 +2132,6 @@ pub(crate) unsafe fn from_untrusted_node_address(candidate: UntrustedNodeAddress
     DomRoot::from_ref(node)
 }
 
-impl<'dom> LayoutDom<'dom, Node> {
-    #[inline]
-    #[expect(unsafe_code)]
-    pub(crate) fn parent_node_ref(self) -> Option<LayoutDom<'dom, Node>> {
-        unsafe { self.unsafe_get().parent_node.get_inner_as_layout() }
-    }
-
-    #[inline]
-    pub(crate) fn type_id_for_layout(self) -> NodeTypeId {
-        self.unsafe_get().type_id()
-    }
-
-    #[inline]
-    pub(crate) fn is_element_for_layout(&self) -> bool {
-        (*self).is::<Element>()
-    }
-
-    pub(crate) fn is_text_node_for_layout(&self) -> bool {
-        matches!(
-            self.type_id_for_layout(),
-            NodeTypeId::CharacterData(CharacterDataTypeId::Text(..))
-        )
-    }
-
-    #[inline]
-    pub(crate) fn composed_parent_node_ref(self) -> Option<LayoutDom<'dom, Node>> {
-        let parent = self.parent_node_ref();
-        if let Some(parent) = parent &&
-            let Some(shadow_root) = parent.downcast::<ShadowRoot>()
-        {
-            return Some(shadow_root.get_host_for_layout().upcast());
-        }
-        parent
-    }
-
-    #[inline]
-    pub(crate) fn traversal_parent(self) -> Option<LayoutDom<'dom, Element>> {
-        if let Some(assigned_slot) = self.assigned_slot_for_layout() {
-            return Some(assigned_slot.upcast());
-        }
-        let parent = self.parent_node_ref()?;
-        if let Some(shadow) = parent.downcast::<ShadowRoot>() {
-            return Some(shadow.get_host_for_layout());
-        };
-        parent.downcast()
-    }
-
-    #[inline]
-    #[expect(unsafe_code)]
-    pub(crate) fn first_child_ref(self) -> Option<LayoutDom<'dom, Node>> {
-        unsafe { self.unsafe_get().first_child.get_inner_as_layout() }
-    }
-
-    #[inline]
-    #[expect(unsafe_code)]
-    pub(crate) fn last_child_ref(self) -> Option<LayoutDom<'dom, Node>> {
-        unsafe { self.unsafe_get().last_child.get_inner_as_layout() }
-    }
-
-    #[inline]
-    #[expect(unsafe_code)]
-    pub(crate) fn prev_sibling_ref(self) -> Option<LayoutDom<'dom, Node>> {
-        unsafe { self.unsafe_get().prev_sibling.get_inner_as_layout() }
-    }
-
-    #[inline]
-    #[expect(unsafe_code)]
-    pub(crate) fn next_sibling_ref(self) -> Option<LayoutDom<'dom, Node>> {
-        unsafe { self.unsafe_get().next_sibling.get_inner_as_layout() }
-    }
-
-    #[inline]
-    #[expect(unsafe_code)]
-    pub(crate) fn owner_doc_for_layout(self) -> LayoutDom<'dom, Document> {
-        unsafe { self.unsafe_get().owner_doc.get_inner_as_layout().unwrap() }
-    }
-
-    #[inline]
-    #[expect(unsafe_code)]
-    pub(crate) fn containing_shadow_root_for_layout(self) -> Option<LayoutDom<'dom, ShadowRoot>> {
-        unsafe {
-            self.unsafe_get()
-                .rare_data
-                .borrow_for_layout()
-                .as_ref()?
-                .containing_shadow_root
-                .as_ref()
-                .map(|sr| sr.to_layout())
-        }
-    }
-
-    #[inline]
-    #[expect(unsafe_code)]
-    pub(crate) fn assigned_slot_for_layout(self) -> Option<LayoutDom<'dom, HTMLSlotElement>> {
-        unsafe {
-            self.unsafe_get()
-                .rare_data
-                .borrow_for_layout()
-                .as_ref()?
-                .slottable_data
-                .assigned_slot
-                .as_ref()
-                .map(|assigned_slot| assigned_slot.to_layout())
-        }
-    }
-
-    // FIXME(nox): get_flag/set_flag (especially the latter) are not safe because
-    // they mutate stuff while values of this type can be used from multiple
-    // threads at once, this should be revisited.
-
-    #[inline]
-    #[expect(unsafe_code)]
-    pub(crate) unsafe fn get_flag(self, flag: NodeFlags) -> bool {
-        (self.unsafe_get()).flags.get().contains(flag)
-    }
-
-    #[inline]
-    #[expect(unsafe_code)]
-    pub(crate) unsafe fn set_flag(self, flag: NodeFlags, value: bool) {
-        let this = self.unsafe_get();
-        let mut flags = (this).flags.get();
-
-        if value {
-            flags.insert(flag);
-        } else {
-            flags.remove(flag);
-        }
-
-        (this).flags.set(flags);
-    }
-
-    #[inline]
-    #[expect(unsafe_code)]
-    pub(crate) fn layout_data(self) -> Option<&'dom GenericLayoutData> {
-        unsafe { self.unsafe_get().layout_data.borrow_for_layout().as_deref() }
-    }
-
-    /// Initialize the style data of this node.
-    ///
-    /// # Safety
-    ///
-    /// This method is unsafe because it modifies the given node during
-    /// layout. Callers should ensure that no other layout thread is
-    /// attempting to read or modify the opaque layout data of this node.
-    #[inline]
-    #[expect(unsafe_code)]
-    pub(crate) unsafe fn initialize_layout_data(self, new_data: Box<GenericLayoutData>) {
-        let data = unsafe { self.unsafe_get().layout_data.borrow_mut_for_layout() };
-        debug_assert!(data.is_none());
-        *data = Some(new_data);
-    }
-
-    /// Clear the style and opaque layout data of this node.
-    ///
-    /// # Safety
-    ///
-    /// This method is unsafe because it modifies the given node during
-    /// layout. Callers should ensure that no other layout thread is
-    /// attempting to read or modify the opaque layout data of this node.
-    #[inline]
-    #[expect(unsafe_code)]
-    pub(crate) unsafe fn clear_layout_data(self) {
-        unsafe {
-            self.unsafe_get().layout_data.borrow_mut_for_layout().take();
-        }
-    }
-
-    /// Whether this element serve as a container of editable text for a text input
-    /// that is implemented as an UA widget.
-    pub(crate) fn is_single_line_text_inner_editor(&self) -> bool {
-        matches!(
-            self.implemented_pseudo_element(),
-            Some(PseudoElement::ServoTextControlInnerEditor)
-        )
-    }
-
-    /// Whether this element serve as a container of any text inside a text input
-    /// that is implemented as an UA widget.
-    pub(crate) fn is_text_container_of_single_line_input(&self) -> bool {
-        let is_single_line_text_inner_placeholder = matches!(
-            self.implemented_pseudo_element(),
-            Some(PseudoElement::Placeholder)
-        );
-        // Currently `::placeholder` is only implemented for single line text input element.
-        debug_assert!(
-            !is_single_line_text_inner_placeholder ||
-                self.containing_shadow_root_for_layout()
-                    .map(|root| root.get_host_for_layout())
-                    .map(|host| host.downcast::<HTMLInputElement>())
-                    .is_some()
-        );
-
-        self.is_single_line_text_inner_editor() || is_single_line_text_inner_placeholder
-    }
-
-    pub(crate) fn text_content(self) -> Cow<'dom, str> {
-        self.downcast::<Text>()
-            .expect("Called LayoutDom::text_content on non-Text node!")
-            .upcast()
-            .data_for_layout()
-            .into()
-    }
-
-    /// Get the selection for the given node. This only works for text nodes that are in
-    /// the shadow DOM of user agent widgets for form controls, specifically for `<input>`
-    /// and `<textarea>`.
-    ///
-    /// As we want to expose the selection on the inner text node of the widget's shadow
-    /// DOM, we must find the shadow root and then access the containing element itself.
-    pub(crate) fn selection(self) -> Option<SharedSelection> {
-        if let Some(input) = self.downcast::<HTMLInputElement>() {
-            return input.selection_for_layout();
-        }
-        if let Some(textarea) = self.downcast::<HTMLTextAreaElement>() {
-            return Some(textarea.selection_for_layout());
-        }
-
-        let shadow_root = self
-            .containing_shadow_root_for_layout()?
-            .get_host_for_layout();
-        if let Some(input) = shadow_root.downcast::<HTMLInputElement>() {
-            return input.selection_for_layout();
-        }
-        shadow_root
-            .downcast::<HTMLTextAreaElement>()
-            .map(|textarea| textarea.selection_for_layout())
-    }
-
-    pub(crate) fn image_url(self) -> Option<ServoUrl> {
-        self.downcast::<HTMLImageElement>()
-            .expect("not an image!")
-            .image_url()
-    }
-
-    pub(crate) fn image_data(self) -> Option<(Option<Image>, Option<ImageMetadata>)> {
-        self.downcast::<HTMLImageElement>().map(|e| e.image_data())
-    }
-
-    pub(crate) fn image_density(self) -> Option<f64> {
-        self.downcast::<HTMLImageElement>()
-            .expect("not an image!")
-            .image_density()
-    }
-
-    pub(crate) fn showing_broken_image_icon(self) -> bool {
-        self.downcast::<HTMLImageElement>()
-            .map(|image_element| image_element.showing_broken_image_icon())
-            .unwrap_or_default()
-    }
-
-    pub(crate) fn canvas_data(self) -> Option<HTMLCanvasData> {
-        self.downcast::<HTMLCanvasElement>()
-            .map(|canvas| canvas.data())
-    }
-
-    pub(crate) fn media_data(self) -> Option<HTMLMediaData> {
-        self.downcast::<HTMLVideoElement>()
-            .map(|media| media.data())
-    }
-
-    pub(crate) fn svg_data(self) -> Option<SVGElementData<'dom>> {
-        self.downcast::<SVGSVGElement>().map(|svg| svg.data())
-    }
-
-    pub(crate) fn iframe_browsing_context_id(self) -> Option<BrowsingContextId> {
-        self.downcast::<HTMLIFrameElement>()
-            .and_then(|iframe_element| iframe_element.browsing_context_id())
-    }
-
-    pub(crate) fn iframe_pipeline_id(self) -> Option<PipelineId> {
-        self.downcast::<HTMLIFrameElement>()
-            .and_then(|iframe_element| iframe_element.pipeline_id())
-    }
-
-    #[expect(unsafe_code)]
-    pub(crate) fn opaque(self) -> OpaqueNode {
-        unsafe { OpaqueNode(self.get_jsobject() as usize) }
-    }
-
-    #[expect(unsafe_code)]
-    pub(crate) fn implemented_pseudo_element(&self) -> Option<PseudoElement> {
-        unsafe {
-            self.unsafe_get()
-                .rare_data
-                .borrow_for_layout()
-                .as_ref()
-                .and_then(|rare_data| rare_data.implemented_pseudo_element)
-        }
-    }
-
-    pub(crate) fn is_in_ua_widget(&self) -> bool {
-        self.unsafe_get().is_in_ua_widget()
-    }
-
-    pub(crate) fn is_root_of_user_agent_widget(&self) -> bool {
-        self.downcast::<Element>().is_some_and(|element| {
-            element
-                .get_shadow_root_for_layout()
-                .is_some_and(|shadow_root| shadow_root.is_user_agent_widget())
-        })
-    }
-}
-
 /// Specifies whether children must be recursively cloned or not.
 #[derive(Clone, Copy, MallocSizeOf, PartialEq)]
 pub(crate) enum CloneChildrenFlag {
@@ -2434,7 +2149,7 @@ impl From<bool> for CloneChildrenFlag {
     }
 }
 
-fn as_uintptr<T>(t: &T) -> uintptr_t {
+pub(super) fn as_uintptr<T>(t: &T) -> uintptr_t {
     t as *const T as uintptr_t
 }
 
@@ -2724,7 +2439,7 @@ impl Node {
     }
 
     /// <https://dom.spec.whatwg.org/#concept-node-insert>
-    fn insert(
+    pub(super) fn insert(
         cx: &mut JSContext,
         node: &Node,
         parent: &Node,
@@ -2973,7 +2688,11 @@ impl Node {
     }
 
     /// <https://dom.spec.whatwg.org/#concept-node-pre-remove>
-    fn pre_remove(cx: &mut JSContext, child: &Node, parent: &Node) -> Fallible<DomRoot<Node>> {
+    pub(super) fn pre_remove(
+        cx: &mut JSContext,
+        child: &Node,
+        parent: &Node,
+    ) -> Fallible<DomRoot<Node>> {
         // Step 1.
         match child.GetParentNode() {
             Some(ref node) if &**node != parent => return Err(Error::NotFound(None)),
@@ -2989,7 +2708,7 @@ impl Node {
     }
 
     /// <https://dom.spec.whatwg.org/#concept-node-remove>
-    fn remove(
+    pub(super) fn remove(
         cx: &mut JSContext,
         node: &Node,
         parent: &Node,
@@ -3604,768 +3323,6 @@ impl Node {
     }
 }
 
-impl NodeMethods<crate::DomTypeHolder> for Node {
-    /// <https://dom.spec.whatwg.org/#dom-node-nodetype>
-    fn NodeType(&self) -> u16 {
-        match self.type_id() {
-            NodeTypeId::Attr => NodeConstants::ATTRIBUTE_NODE,
-            NodeTypeId::CharacterData(CharacterDataTypeId::Text(TextTypeId::Text)) => {
-                NodeConstants::TEXT_NODE
-            },
-            NodeTypeId::CharacterData(CharacterDataTypeId::Text(TextTypeId::CDATASection)) => {
-                NodeConstants::CDATA_SECTION_NODE
-            },
-            NodeTypeId::CharacterData(CharacterDataTypeId::ProcessingInstruction) => {
-                NodeConstants::PROCESSING_INSTRUCTION_NODE
-            },
-            NodeTypeId::CharacterData(CharacterDataTypeId::Comment) => NodeConstants::COMMENT_NODE,
-            NodeTypeId::Document(_) => NodeConstants::DOCUMENT_NODE,
-            NodeTypeId::DocumentType => NodeConstants::DOCUMENT_TYPE_NODE,
-            NodeTypeId::DocumentFragment(_) => NodeConstants::DOCUMENT_FRAGMENT_NODE,
-            NodeTypeId::Element(_) => NodeConstants::ELEMENT_NODE,
-        }
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-nodename>
-    fn NodeName(&self) -> DOMString {
-        match self.type_id() {
-            NodeTypeId::Attr => self.downcast::<Attr>().unwrap().qualified_name(),
-            NodeTypeId::Element(..) => self.downcast::<Element>().unwrap().TagName(),
-            NodeTypeId::CharacterData(CharacterDataTypeId::Text(TextTypeId::Text)) => {
-                DOMString::from("#text")
-            },
-            NodeTypeId::CharacterData(CharacterDataTypeId::Text(TextTypeId::CDATASection)) => {
-                DOMString::from("#cdata-section")
-            },
-            NodeTypeId::CharacterData(CharacterDataTypeId::ProcessingInstruction) => {
-                self.downcast::<ProcessingInstruction>().unwrap().Target()
-            },
-            NodeTypeId::CharacterData(CharacterDataTypeId::Comment) => DOMString::from("#comment"),
-            NodeTypeId::DocumentType => self.downcast::<DocumentType>().unwrap().name().clone(),
-            NodeTypeId::DocumentFragment(_) => DOMString::from("#document-fragment"),
-            NodeTypeId::Document(_) => DOMString::from("#document"),
-        }
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-baseuri>
-    fn BaseURI(&self) -> USVString {
-        USVString(String::from(self.owner_doc().base_url().as_str()))
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-isconnected>
-    fn IsConnected(&self) -> bool {
-        self.is_connected()
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-ownerdocument>
-    fn GetOwnerDocument(&self) -> Option<DomRoot<Document>> {
-        match self.type_id() {
-            NodeTypeId::Document(_) => None,
-            _ => Some(self.owner_doc()),
-        }
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-getrootnode>
-    fn GetRootNode(&self, options: &GetRootNodeOptions) -> DomRoot<Node> {
-        if !options.composed &&
-            let Some(shadow_root) = self.containing_shadow_root()
-        {
-            return DomRoot::upcast(shadow_root);
-        }
-
-        if self.is_connected() {
-            DomRoot::from_ref(self.owner_doc().upcast::<Node>())
-        } else {
-            self.inclusive_ancestors(ShadowIncluding::Yes)
-                .last()
-                .unwrap()
-        }
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-parentnode>
-    fn GetParentNode(&self) -> Option<DomRoot<Node>> {
-        self.parent_node.get()
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-parentelement>
-    fn GetParentElement(&self) -> Option<DomRoot<Element>> {
-        self.GetParentNode().and_then(DomRoot::downcast)
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-haschildnodes>
-    fn HasChildNodes(&self) -> bool {
-        self.first_child.get().is_some()
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-childnodes>
-    fn ChildNodes(&self, cx: &mut JSContext) -> DomRoot<NodeList> {
-        if let Some(list) = self.ensure_rare_data().child_list.get() {
-            return list;
-        }
-
-        let doc = self.owner_doc();
-        let window = doc.window();
-        let list = NodeList::new_child_list(window, self, CanGc::from_cx(cx));
-        self.ensure_rare_data().child_list.set(Some(&list));
-        list
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-firstchild>
-    fn GetFirstChild(&self) -> Option<DomRoot<Node>> {
-        self.first_child.get()
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-lastchild>
-    fn GetLastChild(&self) -> Option<DomRoot<Node>> {
-        self.last_child.get()
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-previoussibling>
-    fn GetPreviousSibling(&self) -> Option<DomRoot<Node>> {
-        self.prev_sibling.get()
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-nextsibling>
-    fn GetNextSibling(&self) -> Option<DomRoot<Node>> {
-        self.next_sibling.get()
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-nodevalue>
-    fn GetNodeValue(&self) -> Option<DOMString> {
-        match self.type_id() {
-            NodeTypeId::Attr => Some(self.downcast::<Attr>().unwrap().Value()),
-            NodeTypeId::CharacterData(_) => {
-                self.downcast::<CharacterData>().map(CharacterData::Data)
-            },
-            _ => None,
-        }
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-nodevalue>
-    fn SetNodeValue(&self, cx: &mut JSContext, val: Option<DOMString>) -> Fallible<()> {
-        match self.type_id() {
-            NodeTypeId::Attr => {
-                let attr = self.downcast::<Attr>().unwrap();
-                attr.SetValue(cx, val.unwrap_or_default())?;
-            },
-            NodeTypeId::CharacterData(_) => {
-                let character_data = self.downcast::<CharacterData>().unwrap();
-                character_data.SetData(cx, val.unwrap_or_default());
-            },
-            _ => {},
-        };
-        Ok(())
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-textcontent>
-    fn GetTextContent(&self) -> Option<DOMString> {
-        match self.type_id() {
-            NodeTypeId::DocumentFragment(_) | NodeTypeId::Element(..) => {
-                let content =
-                    Node::collect_text_contents(self.traverse_preorder(ShadowIncluding::No));
-                Some(content)
-            },
-            NodeTypeId::Attr => Some(self.downcast::<Attr>().unwrap().Value()),
-            NodeTypeId::CharacterData(..) => {
-                let characterdata = self.downcast::<CharacterData>().unwrap();
-                Some(characterdata.Data())
-            },
-            NodeTypeId::DocumentType | NodeTypeId::Document(_) => None,
-        }
-    }
-
-    /// <https://dom.spec.whatwg.org/#set-text-content>
-    fn SetTextContent(&self, cx: &mut JSContext, value: Option<DOMString>) -> Fallible<()> {
-        match self.type_id() {
-            NodeTypeId::DocumentFragment(_) | NodeTypeId::Element(..) => {
-                self.set_text_content_for_element(cx, value);
-            },
-            NodeTypeId::Attr => {
-                let attr = self.downcast::<Attr>().unwrap();
-                attr.SetValue(cx, value.unwrap_or_default())?;
-            },
-            NodeTypeId::CharacterData(..) => {
-                let characterdata = self.downcast::<CharacterData>().unwrap();
-                characterdata.SetData(cx, value.unwrap_or_default());
-            },
-            NodeTypeId::DocumentType | NodeTypeId::Document(_) => {},
-        };
-        Ok(())
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-insertbefore>
-    fn InsertBefore(
-        &self,
-        cx: &mut JSContext,
-        node: &Node,
-        child: Option<&Node>,
-    ) -> Fallible<DomRoot<Node>> {
-        Node::pre_insert(cx, node, self, child)
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-appendchild>
-    fn AppendChild(&self, cx: &mut JSContext, node: &Node) -> Fallible<DomRoot<Node>> {
-        Node::pre_insert(cx, node, self, None)
-    }
-
-    /// <https://dom.spec.whatwg.org/#concept-node-replace>
-    fn ReplaceChild(
-        &self,
-        cx: &mut JSContext,
-        node: &Node,
-        child: &Node,
-    ) -> Fallible<DomRoot<Node>> {
-        // Step 1. If parent is not a Document, DocumentFragment, or Element node,
-        // then throw a "HierarchyRequestError" DOMException.
-        match self.type_id() {
-            NodeTypeId::Document(_) | NodeTypeId::DocumentFragment(_) | NodeTypeId::Element(..) => {
-            },
-            _ => return Err(Error::HierarchyRequest(None)),
-        }
-
-        // Step 2. If node is a host-including inclusive ancestor of parent,
-        // then throw a "HierarchyRequestError" DOMException.
-        if node.is_inclusive_ancestor_of(self) {
-            return Err(Error::HierarchyRequest(None));
-        }
-
-        // Step 3. If child’s parent is not parent, then throw a "NotFoundError" DOMException.
-        if !self.is_parent_of(child) {
-            return Err(Error::NotFound(None));
-        }
-
-        // Step 4. If node is not a DocumentFragment, DocumentType, Element, or CharacterData node,
-        // then throw a "HierarchyRequestError" DOMException.
-        // Step 5. If either node is a Text node and parent is a document,
-        // or node is a doctype and parent is not a document, then throw a "HierarchyRequestError" DOMException.
-        match node.type_id() {
-            NodeTypeId::CharacterData(CharacterDataTypeId::Text(_)) if self.is::<Document>() => {
-                return Err(Error::HierarchyRequest(None));
-            },
-            NodeTypeId::DocumentType if !self.is::<Document>() => {
-                return Err(Error::HierarchyRequest(None));
-            },
-            NodeTypeId::Document(_) | NodeTypeId::Attr => {
-                return Err(Error::HierarchyRequest(None));
-            },
-            _ => (),
-        }
-
-        // Step 6. If parent is a document, and any of the statements below, switched on the interface node implements,
-        // are true, then throw a "HierarchyRequestError" DOMException.
-        if self.is::<Document>() {
-            match node.type_id() {
-                // Step 6.1
-                NodeTypeId::DocumentFragment(_) => {
-                    // Step 6.1.1(b)
-                    if node.children_unrooted(cx.no_gc()).any(|c| c.is::<Text>()) {
-                        return Err(Error::HierarchyRequest(None));
-                    }
-                    match node.child_elements_unrooted(cx.no_gc()).count() {
-                        0 => (),
-                        // Step 6.1.2
-                        1 => {
-                            if self
-                                .child_elements_unrooted(cx.no_gc())
-                                .any(|c| c.upcast::<Node>() != child)
-                            {
-                                return Err(Error::HierarchyRequest(None));
-                            }
-                            if child.following_siblings().any(|child| child.is_doctype()) {
-                                return Err(Error::HierarchyRequest(None));
-                            }
-                        },
-                        // Step 6.1.1(a)
-                        _ => return Err(Error::HierarchyRequest(None)),
-                    }
-                },
-                // Step 6.2
-                NodeTypeId::Element(..) => {
-                    if self
-                        .child_elements_unrooted(cx.no_gc())
-                        .any(|c| c.upcast::<Node>() != child)
-                    {
-                        return Err(Error::HierarchyRequest(None));
-                    }
-                    if child.following_siblings().any(|child| child.is_doctype()) {
-                        return Err(Error::HierarchyRequest(None));
-                    }
-                },
-                // Step 6.3
-                NodeTypeId::DocumentType => {
-                    if self
-                        .children_unrooted(cx.no_gc())
-                        .any(|c| c.is_doctype() && *c != child)
-                    {
-                        return Err(Error::HierarchyRequest(None));
-                    }
-                    if self
-                        .children_unrooted(cx.no_gc())
-                        .take_while(|c| **c != child)
-                        .any(|c| c.is::<Element>())
-                    {
-                        return Err(Error::HierarchyRequest(None));
-                    }
-                },
-                NodeTypeId::CharacterData(..) => (),
-                // Because Document and Attr should already throw `HierarchyRequest`
-                // error, both of them are unreachable here.
-                NodeTypeId::Document(_) => unreachable!(),
-                NodeTypeId::Attr => unreachable!(),
-            }
-        }
-
-        // Step 7. Let referenceChild be child’s next sibling.
-        // Step 8. If referenceChild is node, then set referenceChild to node’s next sibling.
-        let child_next_sibling = child.GetNextSibling();
-        let node_next_sibling = node.GetNextSibling();
-        let reference_child = if child_next_sibling.as_deref() == Some(node) {
-            node_next_sibling.as_deref()
-        } else {
-            child_next_sibling.as_deref()
-        };
-
-        // Step 9. Let previousSibling be child’s previous sibling.
-        let previous_sibling = child.GetPreviousSibling();
-
-        // NOTE: All existing browsers assume that adoption is performed here, which does not follow the DOM spec.
-        // However, if we follow the spec and delay adoption to inside `Node::insert()`, then the mutation records will
-        // be different, and we will fail WPT dom/nodes/MutationObserver-childList.html.
-        let document = self.owner_document();
-        Node::adopt(cx, node, &document);
-
-        // Step 10. Let removedNodes be the empty set.
-        // Step 11. If child’s parent is non-null:
-        //     1. Set removedNodes to « child ».
-        //     2. Remove child with the suppress observers flag set.
-        let removed_child = if node != child {
-            // Step 11.
-            Node::remove(cx, child, self, SuppressObserver::Suppressed);
-            Some(child)
-        } else {
-            None
-        };
-
-        // Step 12. Let nodes be node’s children if node is a DocumentFragment node; otherwise « node ».
-        rooted_vec!(let mut nodes);
-        let nodes = if node.type_id() ==
-            NodeTypeId::DocumentFragment(DocumentFragmentTypeId::DocumentFragment) ||
-            node.type_id() == NodeTypeId::DocumentFragment(DocumentFragmentTypeId::ShadowRoot)
-        {
-            nodes.extend(node.children().map(|node| Dom::from_ref(&*node)));
-            nodes.r()
-        } else {
-            from_ref(&node)
-        };
-
-        // Step 13. Insert node into parent before referenceChild with the suppress observers flag set.
-        Node::insert(
-            cx,
-            node,
-            self,
-            reference_child,
-            SuppressObserver::Suppressed,
-        );
-
-        vtable_for(self).children_changed(
-            cx,
-            &ChildrenMutation::replace(
-                previous_sibling.as_deref(),
-                &removed_child,
-                nodes,
-                reference_child,
-            ),
-        );
-
-        // Step 14. Queue a tree mutation record for parent with nodes, removedNodes,
-        // previousSibling, and referenceChild.
-        let removed = removed_child.map(|r| [r]);
-        let mutation = LazyCell::new(|| Mutation::ChildList {
-            added: Some(nodes),
-            removed: removed.as_ref().map(|r| &r[..]),
-            prev: previous_sibling.as_deref(),
-            next: reference_child,
-        });
-
-        MutationObserver::queue_a_mutation_record(self, mutation);
-
-        // Step 15. Return child.
-        Ok(DomRoot::from_ref(child))
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-removechild>
-    fn RemoveChild(&self, cx: &mut JSContext, node: &Node) -> Fallible<DomRoot<Node>> {
-        Node::pre_remove(cx, node, self)
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-normalize>
-    fn Normalize(&self, cx: &mut JSContext) {
-        let mut children = self.children().enumerate().peekable();
-        while let Some((_, node)) = children.next() {
-            if let Some(text) = node.downcast::<Text>() {
-                if text.is::<CDATASection>() {
-                    continue;
-                }
-                let cdata = text.upcast::<CharacterData>();
-                let mut length = cdata.Length();
-                if length == 0 {
-                    Node::remove(cx, &node, self, SuppressObserver::Unsuppressed);
-                    continue;
-                }
-                while children.peek().is_some_and(|(_, sibling)| {
-                    sibling.is::<Text>() && !sibling.is::<CDATASection>()
-                }) {
-                    let (index, sibling) = children.next().unwrap();
-                    sibling
-                        .ranges()
-                        .drain_to_preceding_text_sibling(&sibling, &node, length);
-                    self.ranges()
-                        .move_to_text_child_at(self, index as u32, &node, length);
-                    let sibling_cdata = sibling.downcast::<CharacterData>().unwrap();
-                    length += sibling_cdata.Length();
-                    cdata.append_data(cx, &sibling_cdata.data());
-                    Node::remove(cx, &sibling, self, SuppressObserver::Unsuppressed);
-                }
-            } else {
-                node.Normalize(cx);
-            }
-        }
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-clonenode>
-    fn CloneNode(&self, cx: &mut JSContext, subtree: bool) -> Fallible<DomRoot<Node>> {
-        // Step 1. If this is a shadow root, then throw a "NotSupportedError" DOMException.
-        if self.is::<ShadowRoot>() {
-            return Err(Error::NotSupported(None));
-        }
-
-        // Step 2. Return the result of cloning a node given this with subtree set to subtree.
-        let result = Node::clone(
-            cx,
-            self,
-            None,
-            if subtree {
-                CloneChildrenFlag::CloneChildren
-            } else {
-                CloneChildrenFlag::DoNotCloneChildren
-            },
-            None,
-        );
-        Ok(result)
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-isequalnode>
-    fn IsEqualNode(&self, maybe_node: Option<&Node>) -> bool {
-        fn is_equal_doctype(node: &Node, other: &Node) -> bool {
-            let doctype = node.downcast::<DocumentType>().unwrap();
-            let other_doctype = other.downcast::<DocumentType>().unwrap();
-            (*doctype.name() == *other_doctype.name()) &&
-                (*doctype.public_id() == *other_doctype.public_id()) &&
-                (*doctype.system_id() == *other_doctype.system_id())
-        }
-        fn is_equal_element(node: &Node, other: &Node) -> bool {
-            let element = node.downcast::<Element>().unwrap();
-            let other_element = other.downcast::<Element>().unwrap();
-            (*element.namespace() == *other_element.namespace()) &&
-                (*element.prefix() == *other_element.prefix()) &&
-                (*element.local_name() == *other_element.local_name()) &&
-                (element.attrs().borrow().len() == other_element.attrs().borrow().len())
-        }
-        fn is_equal_processinginstruction(node: &Node, other: &Node) -> bool {
-            let pi = node.downcast::<ProcessingInstruction>().unwrap();
-            let other_pi = other.downcast::<ProcessingInstruction>().unwrap();
-            (*pi.target() == *other_pi.target()) &&
-                (*pi.upcast::<CharacterData>().data() ==
-                    *other_pi.upcast::<CharacterData>().data())
-        }
-        fn is_equal_characterdata(node: &Node, other: &Node) -> bool {
-            let characterdata = node.downcast::<CharacterData>().unwrap();
-            let other_characterdata = other.downcast::<CharacterData>().unwrap();
-            *characterdata.data() == *other_characterdata.data()
-        }
-        fn is_equal_attr(node: &Node, other: &Node) -> bool {
-            let attr = node.downcast::<Attr>().unwrap();
-            let other_attr = other.downcast::<Attr>().unwrap();
-            (*attr.namespace() == *other_attr.namespace()) &&
-                (attr.local_name() == other_attr.local_name()) &&
-                (**attr.value() == **other_attr.value())
-        }
-        fn is_equal_element_attrs(node: &Node, other: &Node) -> bool {
-            let element = node.downcast::<Element>().unwrap();
-            let other_element = other.downcast::<Element>().unwrap();
-            assert!(element.attrs().borrow().len() == other_element.attrs().borrow().len());
-            element.attrs().borrow().iter().all(|attr| {
-                other_element.attrs().borrow().iter().any(|other_attr| {
-                    (*attr.namespace() == *other_attr.namespace()) &&
-                        (attr.local_name() == other_attr.local_name()) &&
-                        (**attr.value() == **other_attr.value())
-                })
-            })
-        }
-
-        fn is_equal_node(this: &Node, node: &Node) -> bool {
-            // Step 2.
-            if this.NodeType() != node.NodeType() {
-                return false;
-            }
-
-            match node.type_id() {
-                // Step 3.
-                NodeTypeId::DocumentType if !is_equal_doctype(this, node) => return false,
-                NodeTypeId::Element(..) if !is_equal_element(this, node) => return false,
-                NodeTypeId::CharacterData(CharacterDataTypeId::ProcessingInstruction)
-                    if !is_equal_processinginstruction(this, node) =>
-                {
-                    return false;
-                },
-                NodeTypeId::CharacterData(CharacterDataTypeId::Text(_)) |
-                NodeTypeId::CharacterData(CharacterDataTypeId::Comment)
-                    if !is_equal_characterdata(this, node) =>
-                {
-                    return false;
-                },
-                // Step 4.
-                NodeTypeId::Element(..) if !is_equal_element_attrs(this, node) => return false,
-                NodeTypeId::Attr if !is_equal_attr(this, node) => return false,
-
-                _ => (),
-            }
-
-            // Step 5.
-            if this.children_count() != node.children_count() {
-                return false;
-            }
-
-            // Step 6.
-            this.children()
-                .zip(node.children())
-                .all(|(child, other_child)| is_equal_node(&child, &other_child))
-        }
-        match maybe_node {
-            // Step 1.
-            None => false,
-            // Step 2-6.
-            Some(node) => is_equal_node(self, node),
-        }
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-issamenode>
-    fn IsSameNode(&self, other_node: Option<&Node>) -> bool {
-        match other_node {
-            Some(node) => self == node,
-            None => false,
-        }
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-comparedocumentposition>
-    fn CompareDocumentPosition(&self, other: &Node) -> u16 {
-        // Step 1. If this is other, then return zero.
-        if self == other {
-            return 0;
-        }
-
-        // Step 2. Let node1 be other and node2 be this.
-        let mut node1 = Some(other);
-        let mut node2 = Some(self);
-
-        // Step 3. Let attr1 and attr2 be null.
-        let mut attr1: Option<&Attr> = None;
-        let mut attr2: Option<&Attr> = None;
-
-        // step 4: spec says to operate on node1 here,
-        // node1 is definitely Some(other) going into this step
-        // The compiler doesn't know the lifetime of attr1.GetOwnerElement
-        // is guaranteed by the lifetime of attr1, so we hold it explicitly
-        let attr1owner;
-        if let Some(a) = other.downcast::<Attr>() {
-            attr1 = Some(a);
-            attr1owner = a.GetOwnerElement();
-            node1 = match attr1owner {
-                Some(ref e) => Some(e.upcast()),
-                None => None,
-            }
-        }
-
-        // step 5.1: spec says to operate on node2 here,
-        // node2 is definitely just Some(self) going into this step
-        let attr2owner;
-        if let Some(a) = self.downcast::<Attr>() {
-            attr2 = Some(a);
-            attr2owner = a.GetOwnerElement();
-            node2 = match attr2owner {
-                Some(ref e) => Some(e.upcast()),
-                None => None,
-            }
-        }
-
-        // Step 5.2
-        // This substep seems lacking in test coverage.
-        // We hit this when comparing two attributes that have the
-        // same owner element.
-        if let Some(node2) = node2 &&
-            Some(node2) == node1 &&
-            let (Some(a1), Some(a2)) = (attr1, attr2)
-        {
-            let attrs = node2.downcast::<Element>().unwrap().attrs();
-            // go through the attrs in order to see if self
-            // or other is first; spec is clear that we
-            // want value-equality, not reference-equality
-            for attr in attrs.borrow().iter() {
-                if (*attr.namespace() == *a1.namespace()) &&
-                    (attr.local_name() == a1.local_name()) &&
-                    (**attr.value() == **a1.value())
-                {
-                    return NodeConstants::DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC +
-                        NodeConstants::DOCUMENT_POSITION_PRECEDING;
-                }
-                if (*attr.namespace() == *a2.namespace()) &&
-                    (attr.local_name() == a2.local_name()) &&
-                    (**attr.value() == **a2.value())
-                {
-                    return NodeConstants::DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC +
-                        NodeConstants::DOCUMENT_POSITION_FOLLOWING;
-                }
-            }
-            // both attrs have node2 as their owner element, so
-            // we can't have left the loop without seeing them
-            unreachable!();
-        }
-
-        // Step 6
-        match (node1, node2) {
-            (None, _) => {
-                // node1 is null
-                NodeConstants::DOCUMENT_POSITION_FOLLOWING +
-                    NodeConstants::DOCUMENT_POSITION_DISCONNECTED +
-                    NodeConstants::DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC
-            },
-            (_, None) => {
-                // node2 is null
-                NodeConstants::DOCUMENT_POSITION_PRECEDING +
-                    NodeConstants::DOCUMENT_POSITION_DISCONNECTED +
-                    NodeConstants::DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC
-            },
-            (Some(node1), Some(node2)) => {
-                // still step 6, testing if node1 and 2 share a root
-                let mut self_and_ancestors = node2
-                    .inclusive_ancestors(ShadowIncluding::No)
-                    .collect::<SmallVec<[_; 20]>>();
-                let mut other_and_ancestors = node1
-                    .inclusive_ancestors(ShadowIncluding::No)
-                    .collect::<SmallVec<[_; 20]>>();
-
-                if self_and_ancestors.last() != other_and_ancestors.last() {
-                    let random = as_uintptr(self_and_ancestors.last().unwrap()) <
-                        as_uintptr(other_and_ancestors.last().unwrap());
-                    let random = if random {
-                        NodeConstants::DOCUMENT_POSITION_FOLLOWING
-                    } else {
-                        NodeConstants::DOCUMENT_POSITION_PRECEDING
-                    };
-
-                    // Disconnected.
-                    return random +
-                        NodeConstants::DOCUMENT_POSITION_DISCONNECTED +
-                        NodeConstants::DOCUMENT_POSITION_IMPLEMENTATION_SPECIFIC;
-                }
-                // steps 7-10
-                let mut parent = self_and_ancestors.pop().unwrap();
-                other_and_ancestors.pop().unwrap();
-
-                let mut current_position =
-                    cmp::min(self_and_ancestors.len(), other_and_ancestors.len());
-
-                while current_position > 0 {
-                    current_position -= 1;
-                    let child_1 = self_and_ancestors.pop().unwrap();
-                    let child_2 = other_and_ancestors.pop().unwrap();
-
-                    if child_1 != child_2 {
-                        for child in parent.children() {
-                            if child == child_1 {
-                                // `other` is following `self`.
-                                return NodeConstants::DOCUMENT_POSITION_FOLLOWING;
-                            }
-                            if child == child_2 {
-                                // `other` is preceding `self`.
-                                return NodeConstants::DOCUMENT_POSITION_PRECEDING;
-                            }
-                        }
-                    }
-
-                    parent = child_1;
-                }
-
-                // We hit the end of one of the parent chains, so one node needs to be
-                // contained in the other.
-                //
-                // If we're the container, return that `other` is contained by us.
-                if self_and_ancestors.len() < other_and_ancestors.len() {
-                    NodeConstants::DOCUMENT_POSITION_FOLLOWING +
-                        NodeConstants::DOCUMENT_POSITION_CONTAINED_BY
-                } else {
-                    NodeConstants::DOCUMENT_POSITION_PRECEDING +
-                        NodeConstants::DOCUMENT_POSITION_CONTAINS
-                }
-            },
-        }
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-contains>
-    fn Contains(&self, maybe_other: Option<&Node>) -> bool {
-        match maybe_other {
-            None => false,
-            Some(other) => self.is_inclusive_ancestor_of(other),
-        }
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-lookupprefix>
-    fn LookupPrefix(&self, namespace: Option<DOMString>) -> Option<DOMString> {
-        let namespace = namespace_from_domstring(namespace);
-
-        // Step 1.
-        if namespace == ns!() {
-            return None;
-        }
-
-        // Step 2.
-        match self.type_id() {
-            NodeTypeId::Element(..) => self.downcast::<Element>().unwrap().lookup_prefix(namespace),
-            NodeTypeId::Document(_) => self
-                .downcast::<Document>()
-                .unwrap()
-                .GetDocumentElement()
-                .and_then(|element| element.lookup_prefix(namespace)),
-            NodeTypeId::DocumentType | NodeTypeId::DocumentFragment(_) => None,
-            NodeTypeId::Attr => self
-                .downcast::<Attr>()
-                .unwrap()
-                .GetOwnerElement()
-                .and_then(|element| element.lookup_prefix(namespace)),
-            _ => self
-                .GetParentElement()
-                .and_then(|element| element.lookup_prefix(namespace)),
-        }
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-lookupnamespaceuri>
-    fn LookupNamespaceURI(&self, prefix: Option<DOMString>) -> Option<DOMString> {
-        // Step 1. If prefix is the empty string, then set it to null.
-        let prefix = prefix.filter(|prefix| !prefix.is_empty());
-
-        // Step 2. Return the result of running locate a namespace for this using prefix.
-        Node::namespace_to_string(Node::locate_namespace(self, prefix))
-    }
-
-    /// <https://dom.spec.whatwg.org/#dom-node-isdefaultnamespace>
-    fn IsDefaultNamespace(&self, namespace: Option<DOMString>) -> bool {
-        // Step 1.
-        let namespace = namespace_from_domstring(namespace);
-        // Steps 2 and 3.
-        Node::locate_namespace(self, None) == namespace
-    }
-}
-
 pub(crate) trait NodeTraits {
     /// Get the [`Document`] that owns this node. Note that this may differ from the
     /// [`Document`] that the node was created in if it was adopted by a different
@@ -4422,7 +3379,7 @@ impl VirtualMethods for Node {
             s.children_changed(cx, mutation);
         }
 
-        if let Some(data) = self.rare_data().as_ref() &&
+        if let Some(data) = self.rare_data.borrow().as_ref() &&
             let Some(list) = data.child_list.get()
         {
             list.as_children_list().children_changed(mutation);
@@ -4497,310 +3454,6 @@ pub(crate) enum NodeDamage {
     Other,
 }
 
-pub(crate) enum ChildrenMutation<'a> {
-    Append {
-        prev: &'a Node,
-        added: &'a [&'a Node],
-    },
-    Insert {
-        prev: &'a Node,
-        added: &'a [&'a Node],
-        next: &'a Node,
-    },
-    Prepend {
-        added: &'a [&'a Node],
-        next: &'a Node,
-    },
-    Replace {
-        prev: Option<&'a Node>,
-        removed: &'a Node,
-        added: &'a [&'a Node],
-        next: Option<&'a Node>,
-    },
-    ReplaceAll {
-        removed: &'a [&'a Node],
-        added: &'a [&'a Node],
-    },
-    /// Mutation for when a Text node's data is modified.
-    /// This doesn't change the structure of the list, which is what the other
-    /// variants' fields are stored for at the moment, so this can just have no
-    /// fields.
-    ChangeText,
-}
-
-impl<'a> ChildrenMutation<'a> {
-    fn insert(
-        prev: Option<&'a Node>,
-        added: &'a [&'a Node],
-        next: Option<&'a Node>,
-    ) -> ChildrenMutation<'a> {
-        match (prev, next) {
-            (None, None) => ChildrenMutation::ReplaceAll {
-                removed: &[],
-                added,
-            },
-            (Some(prev), None) => ChildrenMutation::Append { prev, added },
-            (None, Some(next)) => ChildrenMutation::Prepend { added, next },
-            (Some(prev), Some(next)) => ChildrenMutation::Insert { prev, added, next },
-        }
-    }
-
-    fn replace(
-        prev: Option<&'a Node>,
-        removed: &'a Option<&'a Node>,
-        added: &'a [&'a Node],
-        next: Option<&'a Node>,
-    ) -> ChildrenMutation<'a> {
-        if let Some(ref removed) = *removed {
-            if let (None, None) = (prev, next) {
-                ChildrenMutation::ReplaceAll {
-                    removed: from_ref(removed),
-                    added,
-                }
-            } else {
-                ChildrenMutation::Replace {
-                    prev,
-                    removed,
-                    added,
-                    next,
-                }
-            }
-        } else {
-            ChildrenMutation::insert(prev, added, next)
-        }
-    }
-
-    fn replace_all(removed: &'a [&'a Node], added: &'a [&'a Node]) -> ChildrenMutation<'a> {
-        ChildrenMutation::ReplaceAll { removed, added }
-    }
-
-    /// Get the child that follows the added or removed children.
-    /// Currently only used when this mutation might force us to
-    /// restyle later children (see HAS_SLOW_SELECTOR_LATER_SIBLINGS and
-    /// Element's implementation of VirtualMethods::children_changed).
-    pub(crate) fn next_child(&self) -> Option<&Node> {
-        match *self {
-            ChildrenMutation::Append { .. } => None,
-            ChildrenMutation::Insert { next, .. } => Some(next),
-            ChildrenMutation::Prepend { next, .. } => Some(next),
-            ChildrenMutation::Replace { next, .. } => next,
-            ChildrenMutation::ReplaceAll { .. } => None,
-            ChildrenMutation::ChangeText => None,
-        }
-    }
-
-    /// If nodes were added or removed at the start or end of a container, return any
-    /// previously-existing child whose ":first-child" or ":last-child" status *may* have changed.
-    ///
-    /// NOTE: This does not check whether the inserted/removed nodes were elements, so in some
-    /// cases it will return a false positive.  This doesn't matter for correctness, because at
-    /// worst the returned element will be restyled unnecessarily.
-    pub(crate) fn modified_edge_element(&self, no_gc: &NoGC) -> Option<DomRoot<Node>> {
-        match *self {
-            // Add/remove at start of container: Return the first following element.
-            ChildrenMutation::Prepend { next, .. } |
-            ChildrenMutation::Replace {
-                prev: None,
-                next: Some(next),
-                ..
-            } => next
-                .inclusively_following_siblings_unrooted(no_gc)
-                .find(|node| node.is::<Element>())
-                .map(|node| node.as_rooted()),
-            // Add/remove at end of container: Return the last preceding element.
-            ChildrenMutation::Append { prev, .. } |
-            ChildrenMutation::Replace {
-                prev: Some(prev),
-                next: None,
-                ..
-            } => prev
-                .inclusively_preceding_siblings_unrooted(no_gc)
-                .find(|node| node.is::<Element>())
-                .map(|node| node.as_rooted()),
-            // Insert or replace in the middle:
-            ChildrenMutation::Insert { prev, next, .. } |
-            ChildrenMutation::Replace {
-                prev: Some(prev),
-                next: Some(next),
-                ..
-            } => {
-                if prev
-                    .inclusively_preceding_siblings_unrooted(no_gc)
-                    .all(|node| !node.is::<Element>())
-                {
-                    // Before the first element: Return the first following element.
-                    next.inclusively_following_siblings_unrooted(no_gc)
-                        .find(|node| node.is::<Element>())
-                        .map(|node| node.as_rooted())
-                } else if next
-                    .inclusively_following_siblings_unrooted(no_gc)
-                    .all(|node| !node.is::<Element>())
-                {
-                    // After the last element: Return the last preceding element.
-                    prev.inclusively_preceding_siblings_unrooted(no_gc)
-                        .find(|node| node.is::<Element>())
-                        .map(|node| node.as_rooted())
-                } else {
-                    None
-                }
-            },
-
-            ChildrenMutation::Replace {
-                prev: None,
-                next: None,
-                ..
-            } => unreachable!(),
-            ChildrenMutation::ReplaceAll { .. } => None,
-            ChildrenMutation::ChangeText => None,
-        }
-    }
-}
-
-/// The context of the binding to tree of a node.
-pub(crate) struct BindContext<'a> {
-    /// The parent of the inclusive ancestor that was inserted.
-    pub(crate) parent: &'a Node,
-
-    /// Whether the tree is connected.
-    ///
-    /// <https://dom.spec.whatwg.org/#connected>
-    pub(crate) tree_connected: bool,
-
-    /// Whether the tree's root is a document.
-    ///
-    /// <https://dom.spec.whatwg.org/#in-a-document-tree>
-    pub(crate) tree_is_in_a_document_tree: bool,
-
-    /// Whether the tree's root is a shadow root
-    pub(crate) tree_is_in_a_shadow_tree: bool,
-
-    /// Whether the root of the subtree that is being bound to the parent is a shadow root.
-    ///
-    /// This implies that all elements whose "bind_to_tree" method are called were already
-    /// in a shadow tree beforehand.
-    pub(crate) is_shadow_tree: IsShadowTree,
-}
-
-#[derive(Debug, Eq, PartialEq)]
-pub(crate) enum IsShadowTree {
-    Yes,
-    No,
-}
-
-impl<'a> BindContext<'a> {
-    /// Create a new `BindContext` value.
-    pub(crate) fn new(parent: &'a Node, is_shadow_tree: IsShadowTree) -> Self {
-        BindContext {
-            parent,
-            tree_connected: parent.is_connected(),
-            tree_is_in_a_document_tree: parent.is_in_a_document_tree(),
-            tree_is_in_a_shadow_tree: parent.is_in_a_shadow_tree(),
-            is_shadow_tree,
-        }
-    }
-
-    /// Return true iff the tree is inside either a document- or a shadow tree.
-    pub(crate) fn is_in_tree(&self) -> bool {
-        self.tree_is_in_a_document_tree || self.tree_is_in_a_shadow_tree
-    }
-}
-
-/// The context of the unbinding from a tree of a node when one of its
-/// inclusive ancestors is removed.
-pub(crate) struct UnbindContext<'a> {
-    /// The index of the inclusive ancestor that was removed.
-    index: Cell<Option<u32>>,
-    /// The parent of the inclusive ancestor that was removed.
-    pub(crate) parent: &'a Node,
-    /// The previous sibling of the inclusive ancestor that was removed.
-    prev_sibling: Option<&'a Node>,
-    /// The next sibling of the inclusive ancestor that was removed.
-    pub(crate) next_sibling: Option<&'a Node>,
-
-    /// Whether the tree is connected.
-    ///
-    /// <https://dom.spec.whatwg.org/#connected>
-    pub(crate) tree_connected: bool,
-
-    /// Whether the tree's root is a document.
-    ///
-    /// <https://dom.spec.whatwg.org/#in-a-document-tree>
-    pub(crate) tree_is_in_a_document_tree: bool,
-
-    /// Whether the tree's root is a shadow root
-    pub(crate) tree_is_in_a_shadow_tree: bool,
-}
-
-impl<'a> UnbindContext<'a> {
-    /// Create a new `UnbindContext` value.
-    pub(crate) fn new(
-        parent: &'a Node,
-        prev_sibling: Option<&'a Node>,
-        next_sibling: Option<&'a Node>,
-        cached_index: Option<u32>,
-    ) -> Self {
-        UnbindContext {
-            index: Cell::new(cached_index),
-            parent,
-            prev_sibling,
-            next_sibling,
-            tree_connected: parent.is_connected(),
-            tree_is_in_a_document_tree: parent.is_in_a_document_tree(),
-            tree_is_in_a_shadow_tree: parent.is_in_a_shadow_tree(),
-        }
-    }
-
-    /// The index of the inclusive ancestor that was removed from the tree.
-    pub(crate) fn index(&self) -> u32 {
-        if let Some(index) = self.index.get() {
-            return index;
-        }
-        let index = self.prev_sibling.map_or(0, |sibling| sibling.index() + 1);
-        self.index.set(Some(index));
-        index
-    }
-}
-
-/// The context of the moving from a tree of a node when one of its
-/// inclusive ancestors is moved.
-pub(crate) struct MoveContext<'a> {
-    /// The index of the inclusive ancestor that was moved.
-    index: Cell<Option<u32>>,
-    /// The old parent, if any, of the inclusive ancestor that was moved.
-    pub(crate) old_parent: Option<&'a Node>,
-    /// The previous sibling of the inclusive ancestor that was moved.
-    prev_sibling: Option<&'a Node>,
-    /// The next sibling of the inclusive ancestor that was moved.
-    pub(crate) next_sibling: Option<&'a Node>,
-}
-
-impl<'a> MoveContext<'a> {
-    /// Create a new `MoveContext` value.
-    pub(crate) fn new(
-        old_parent: Option<&'a Node>,
-        prev_sibling: Option<&'a Node>,
-        next_sibling: Option<&'a Node>,
-        cached_index: Option<u32>,
-    ) -> Self {
-        MoveContext {
-            index: Cell::new(cached_index),
-            old_parent,
-            prev_sibling,
-            next_sibling,
-        }
-    }
-
-    /// The index of the inclusive ancestor that was moved from the tree.
-    pub(crate) fn index(&self) -> u32 {
-        if let Some(index) = self.index.get() {
-            return index;
-        }
-        let index = self.prev_sibling.map_or(0, |sibling| sibling.index() + 1);
-        self.index.set(Some(index));
-        index
-    }
-}
-
 /// A node's unique ID, for devtools.
 pub(crate) struct UniqueId {
     cell: UnsafeCell<Option<Box<Uuid>>>,
@@ -4821,7 +3474,7 @@ impl MallocSizeOf for UniqueId {
 
 impl UniqueId {
     /// Create a new `UniqueId` value. The underlying `Uuid` is lazily created.
-    fn new() -> UniqueId {
+    pub(super) fn new() -> UniqueId {
         UniqueId {
             cell: UnsafeCell::new(None),
         }
@@ -4829,103 +3482,13 @@ impl UniqueId {
 
     /// The Uuid of that unique ID.
     #[expect(unsafe_code)]
-    fn borrow(&self) -> &Uuid {
+    pub(super) fn borrow(&self) -> &Uuid {
         unsafe {
             let ptr = self.cell.get();
             if (*ptr).is_none() {
                 *ptr = Some(Box::new(Uuid::new_v4()));
             }
             (*ptr).as_ref().unwrap()
-        }
-    }
-}
-
-pub(crate) struct NodeTypeIdWrapper(pub(crate) NodeTypeId);
-
-impl From<NodeTypeIdWrapper> for LayoutNodeType {
-    #[inline(always)]
-    fn from(node_type: NodeTypeIdWrapper) -> LayoutNodeType {
-        match node_type.0 {
-            NodeTypeId::Element(e) => LayoutNodeType::Element(ElementTypeIdWrapper(e).into()),
-            NodeTypeId::CharacterData(CharacterDataTypeId::Text(_)) => LayoutNodeType::Text,
-            x => unreachable!("Layout should not traverse nodes of type {:?}", x),
-        }
-    }
-}
-
-struct ElementTypeIdWrapper(ElementTypeId);
-
-impl From<ElementTypeIdWrapper> for LayoutElementType {
-    #[inline(always)]
-    fn from(element_type: ElementTypeIdWrapper) -> LayoutElementType {
-        match element_type.0 {
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLBodyElement) => {
-                LayoutElementType::HTMLBodyElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLBRElement) => {
-                LayoutElementType::HTMLBRElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLCanvasElement) => {
-                LayoutElementType::HTMLCanvasElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLHtmlElement) => {
-                LayoutElementType::HTMLHtmlElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLIFrameElement) => {
-                LayoutElementType::HTMLIFrameElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLImageElement) => {
-                LayoutElementType::HTMLImageElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLMediaElement(_)) => {
-                LayoutElementType::HTMLMediaElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLInputElement) => {
-                LayoutElementType::HTMLInputElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLOptGroupElement) => {
-                LayoutElementType::HTMLOptGroupElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLOptionElement) => {
-                LayoutElementType::HTMLOptionElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLObjectElement) => {
-                LayoutElementType::HTMLObjectElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLParagraphElement) => {
-                LayoutElementType::HTMLParagraphElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLPreElement) => {
-                LayoutElementType::HTMLPreElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLSelectElement) => {
-                LayoutElementType::HTMLSelectElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLTableCellElement) => {
-                LayoutElementType::HTMLTableCellElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLTableColElement) => {
-                LayoutElementType::HTMLTableColElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLTableElement) => {
-                LayoutElementType::HTMLTableElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLTableRowElement) => {
-                LayoutElementType::HTMLTableRowElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLTableSectionElement) => {
-                LayoutElementType::HTMLTableSectionElement
-            },
-            ElementTypeId::HTMLElement(HTMLElementTypeId::HTMLTextAreaElement) => {
-                LayoutElementType::HTMLTextAreaElement
-            },
-            ElementTypeId::SVGElement(SVGElementTypeId::SVGGraphicsElement(
-                SVGGraphicsElementTypeId::SVGImageElement,
-            )) => LayoutElementType::SVGImageElement,
-            ElementTypeId::SVGElement(SVGElementTypeId::SVGGraphicsElement(
-                SVGGraphicsElementTypeId::SVGSVGElement,
-            )) => LayoutElementType::SVGSVGElement,
-            _ => LayoutElementType::Element,
         }
     }
 }
